+++ main_service/core/models/post.go
@@ -6,20 +6,19 @@ type PostContent struct {
 	Content string `json:"content" binding:"required"`
 }
 
-// TODO: is needed?
 type PostFull struct {
-	PostId          uint32    `json:"post_id" binding:"required"`
-	AuthorId        uint32    `json:"author_id" binding:"required"`
+	PostId          uint64    `json:"post_id" binding:"required"`
+	AuthorId        uint64    `json:"author_id" binding:"required"`
 	Content         string    `json:"content" binding:"required"`
 	CreateTimestamp time.Time `json:"create_timestamp" binding:"required"`
 	UpdateTimestamp time.Time `json:"update_timestamp" binding:"required"`
 }
 
 type Pagination struct {
-	Offset *uint32 `json:"offset" binding:"required"`
+	Offset *uint64 `json:"offset" binding:"required"`
 	Limit  *uint32 `json:"limit" binding:"required"`
 }
 
 type PostId struct {
-	PostId uint32 `uri:"id" binding:"required"`
+	PostId uint64 `uri:"id" binding:"required"`
 }
